fix: revert `Integer` mapping to `float` to avoid Perspective Int32 truncation
Perspective.js v2.10 uses Int32 internally for its `integer` type,
which silently truncates values above 2^31 (~2.1B). Since the Python
bridge classifies `BigInteger` at the 2^53-1 threshold, values between
2^31 and 2^53-1 would be corrupted when mapped as `integer`.

Reverting `Integer` back to `float` preserves all values up to 2^53-1
with correct numeric sorting. The `BigInteger` → `string` mapping
remains for values exceeding 2^53-1.

Addresses Sentry review feedback on #1851.

+++ webview_panels/src/modules/queryPanel/components/perspective/PerspectiveViewer.tsx
@@ -65,7 +65,7 @@ const PerspectiveViewer = ({
       case "Text":
         return "string";
       case "Integer":
-        return "integer";
+        return "float";
       case "BigInteger":
         return "string";
       case "Number":
